create utils module, move funcs over
